Avoiding changes on contacts not belonged to current user

+++ delete.php
@@ -5,7 +5,7 @@ require "session.php";
 
 $id = $_GET["id"];
 
-$statement = $connection->prepare("SELECT * FROM contacts WHERE id = :id;");
+$statement = $connection->prepare("SELECT * FROM contacts WHERE id = :id LIMIT 1;");
 $statement->bindParam(":id", $id);
 $statement->execute();
 
@@ -14,6 +14,14 @@ if ($statement -> rowCount() == 0) {
   echo("<h1>HTTP 404 NOT FOUND :(");
   return;
 }
+
+$contact = $statement -> fetch(PDO::FETCH_ASSOC);
+
+if ($contact["userid"] !== $_SESSION["user"]["id"]) {
+  http_response_code(403);
+  echo ("<h1>UNAUTHORIZED :/");
+  return;
+} 
 
 $statement = $connection->prepare("DELETE FROM contacts WHERE id = :id;");
 $statement->bindParam(":id", $id);
